requirements.txt
djangorestframework
django
django-cors
drf-nested-routers